scaling-controller/requirements.txt
docker==7.0.0
redis==5.0.1
python-dotenv==1.0.0 